adjust mysql/mssql/postgres
