+++ reader/config/config.go
@@ -426,12 +426,13 @@ var ModeKeyOptions = map[string][]Option{
 		},
 		{
 			KeyName:      KeyMysqlDataBase,
+			Required:     true,
 			ChooseOnly:   false,
 			Placeholder:  "<database>，默认读取所有用户数据库",
 			DefaultNoUse: false,
 			Default:      "",
 			Description:  "数据库名称(mysql_database)",
-			ToolTip:      `mysql数据库名称，不填默认读取所有用户数据库`,
+			ToolTip:      `mysql数据库名称，支持魔法变量，*代表所有数据表`,
 		},
 		{
 			KeyName:      KeyMysqlSQL,
@@ -441,7 +442,17 @@ var ModeKeyOptions = map[string][]Option{
 			Placeholder:  "select * from <table>;",
 			DefaultNoUse: true,
 			Description:  "数据查询语句(mysql_sql)",
-			ToolTip:      "填写要执行的sql语句",
+			ToolTip:      "填写要执行的sql语句，不为空则数据库表名字段不生效",
+		},
+		{
+			KeyName:       KyeMysqlTable,
+			ChooseOnly:    false,
+			Placeholder:   "<table>",
+			DefaultNoUse:  true,
+			Default:       "",
+			Description:   "数据库表名(mysql_table)",
+			ToolTip:       "mysql数据库表名，支持魔法变量，*代表所有数据表，若数据库查询语句不为空，则该字段不生效，否则，自动导入匹配数据库名和表名的数据",
+			AdvanceDepend: KeyMysqlHistoryAll,
 		},
 		{
 			KeyName:    KeyMysqlEncoding,
@@ -501,21 +512,14 @@ var ModeKeyOptions = map[string][]Option{
 		},
 		{
 			KeyName:       KeyMysqlHistoryAll,
+			Advance:       true,
 			Element:       Radio,
 			ChooseOnly:    true,
 			ChooseOptions: []interface{}{"true", "false"},
 			Default:       "false",
 			DefaultNoUse:  false,
-			Description:   "导入历史数据(history_all)",
-			ToolTip:       "帮助导入符合要求的所有历史数据",
-		},
-		{
-			KeyName:      KyeMysqlTable,
-			ChooseOnly:   false,
-			Placeholder:  "<table>",
-			DefaultNoUse: true,
-			Default:      "",
-			Description:  "数据库表名(mysql_table)",
+			Description:   "导入匹配的数据(history_all)",
+			ToolTip:       "帮助导入匹配和小于(如果使用了魔法变量)数据库名和表名的所有数据",
 		},
 		OptionSQLSchema,
 		OptionMagicLagDuration,
