+++ reader/sql/dataconvert.go
@@ -1,7 +1,6 @@
 package sql
 
 import (
-	"database/sql"
 	"encoding/binary"
 	"errors"
 	"fmt"
@@ -11,10 +10,10 @@ import (
 
 	"github.com/qiniu/log"
 
-	. "github.com/qiniu/logkit/utils/models"
+	"github.com/qiniu/logkit/utils/models"
 )
 
-func convertLong(v interface{}) (int64, error) {
+func ConvertLong(v interface{}) (int64, error) {
 	dpv := reflect.ValueOf(v)
 	if dpv.Kind() != reflect.Ptr {
 		return 0, errors.New("scanArgs not a pointer")
@@ -80,7 +79,7 @@ func convertLong(v interface{}) (int64, error) {
 	return 0, fmt.Errorf("%v type can not convert to int", dv.Kind())
 }
 
-func convertFloat(v interface{}) (float64, error) {
+func ConvertFloat(v interface{}) (float64, error) {
 	dpv := reflect.ValueOf(v)
 	if dpv.Kind() != reflect.Ptr {
 		return 0, errors.New("scanArgs not a pointer")
@@ -150,7 +149,7 @@ func convertFloat(v interface{}) (float64, error) {
 	return 0, fmt.Errorf("%v type can not convert to int", dv.Kind())
 }
 
-func convertDate(v interface{}) (time.Time, error) {
+func ConvertDate(v interface{}) (time.Time, error) {
 	dpv := reflect.ValueOf(v)
 	if dpv.Kind() != reflect.Ptr {
 		return time.Time{}, errors.New("scanArgs not a pointer")
@@ -176,7 +175,7 @@ func convertDate(v interface{}) (time.Time, error) {
 	return time.Time{}, fmt.Errorf("%v type can not convert to string", dv.Kind())
 }
 
-func convertString(v interface{}) (string, error) {
+func ConvertString(v interface{}) (string, error) {
 	dpv := reflect.ValueOf(v)
 	if dpv.Kind() != reflect.Ptr {
 		return "", errors.New("scanArgs not a pointer")
@@ -242,7 +241,7 @@ func convertString(v interface{}) (string, error) {
 	return "", fmt.Errorf("%v type can not convert to string", dv.Kind())
 }
 
-func convertBool(v interface{}) (bool, error) {
+func ConvertBool(v interface{}) (bool, error) {
 	dpv := reflect.ValueOf(v)
 	if dpv.Kind() != reflect.Ptr {
 		return false, errors.New("scanArgs not a pointer")
@@ -265,127 +264,110 @@ func convertBool(v interface{}) (bool, error) {
 	return false, fmt.Errorf("%v type can not convert to Bool", dv.Kind())
 }
 
-func (r *Reader) getData(rows *sql.Rows, scanArgs []interface{}, columns []string, nochiced []bool) (Data, int64) {
-	// get RawBytes from data
-	err := rows.Scan(scanArgs...)
-	if err != nil {
-		err = fmt.Errorf("runner[%v] %v scan rows error %v", r.meta.RunnerName, r.Name(), err)
-		log.Error(err)
-		r.sendError(err)
-		return nil, 0
+func ConvertScanArgs(data models.Data, scanArg interface{}, column, runnerName, name string,
+	nochiced bool, schemas map[string]string) (int64, error) {
+	var (
+		bytes int64
+		err   error
+	)
+	vtype, ok := schemas[column]
+	if !ok {
+		vtype = "unknown"
 	}
-
-	var totalBytes int64
-	data := make(Data, len(scanArgs))
-	for i := 0; i < len(scanArgs); i++ {
-		var bytes int64
-		vtype, ok := r.schemas[columns[i]]
-		if !ok {
-			vtype = "unknown"
+	switch vtype {
+	case "long":
+		val, serr := ConvertLong(scanArg)
+		if serr != nil {
+			err = fmt.Errorf("runner[%v] %v convertLong for %v (%v) error %v, this key will be ignored", runnerName, name, column, scanArg, serr)
+			log.Error(err)
+		} else {
+			data[column] = val
+			bytes = 8
 		}
-		switch vtype {
-		case "long":
-			val, serr := convertLong(scanArgs[i])
-			if serr != nil {
-				serr = fmt.Errorf("runner[%v] %v convertLong for %v (%v) error %v, this key will be ignored", r.meta.RunnerName, r.Name(), columns[i], scanArgs[i], serr)
-				log.Error(serr)
-				r.sendError(serr)
-			} else {
-				data[columns[i]] = val
+	case "float":
+		val, serr := ConvertFloat(scanArg)
+		if serr != nil {
+			err = fmt.Errorf("runner[%v] %v convertFloat for %v (%v) error %v, this key will be ignored", runnerName, name, column, scanArg, serr)
+			log.Error(err)
+		} else {
+			data[column] = val
+			bytes = 8
+		}
+	case "string":
+		val, serr := ConvertString(scanArg)
+		if serr != nil {
+			err = fmt.Errorf("runner[%v] %v convertString for %v (%v) error %v, this key will be ignored", runnerName, name, column, scanArg, serr)
+			log.Error(err)
+		} else {
+			data[column] = val
+			bytes = int64(len(val))
+		}
+	case "bool":
+		val, serr := ConvertBool(scanArg)
+		if serr != nil {
+			err = fmt.Errorf("runner[%v] %v convertBool for %v (%v) error %v, this key will be ignored", runnerName, name, column, scanArg, serr)
+			log.Error(err)
+		} else {
+			data[column] = val
+			bytes = 4
+		}
+	case "date":
+		val, serr := ConvertDate(scanArg)
+		if serr != nil {
+			err = fmt.Errorf("runner[%v] %v convertDate for %v (%v) error %v, this key will be ignored", runnerName, name, column, scanArg, serr)
+			log.Error(err)
+		} else {
+			data[column] = val
+			bytes = 20
+		}
+	default:
+		dealed := false
+		if !nochiced {
+			dealed = true
+			switch d := scanArg.(type) {
+			case *string:
+				data[column] = *d
+				bytes = int64(len(*d))
+			case *[]byte:
+				data[column] = string(*d)
+				bytes = int64(len(*d))
+			case *bool:
+				data[column] = *d
+				bytes = 1
+			case int64:
+				data[column] = d
 				bytes = 8
-			}
-		case "float":
-			val, serr := convertFloat(scanArgs[i])
-			if serr != nil {
-				serr = fmt.Errorf("runner[%v] %v convertFloat for %v (%v) error %v, this key will be ignored", r.meta.RunnerName, r.Name(), columns[i], scanArgs[i], serr)
-				log.Error(serr)
-				r.sendError(serr)
-			} else {
-				data[columns[i]] = val
+			case *int64:
+				data[column] = *d
 				bytes = 8
+			case float64:
+				data[column] = d
+				bytes = 8
+			case *float64:
+				data[column] = *d
+				bytes = 8
+			case uint64:
+				data[column] = d
+				bytes = 8
+			case *uint64:
+				data[column] = *d
+				bytes = 8
+			case *interface{}:
+				dealed = false
+			default:
+				dealed = false
 			}
-		case "string":
-			val, serr := convertString(scanArgs[i])
+		}
+		if !dealed {
+			val, serr := ConvertString(scanArg)
 			if serr != nil {
-				serr = fmt.Errorf("runner[%v] %v convertString for %v (%v) error %v, this key will be ignored", r.meta.RunnerName, r.Name(), columns[i], scanArgs[i], serr)
-				log.Error(serr)
-				r.sendError(serr)
+				err = fmt.Errorf("runner[%v] %v convertString for %v (%v) error %v, this key will be ignored", runnerName, name, column, scanArg, serr)
+				log.Error(err)
 			} else {
-				data[columns[i]] = val
+				data[column] = val
 				bytes = int64(len(val))
 			}
-		case "bool":
-			val, serr := convertBool(scanArgs[i])
-			if serr != nil {
-				serr = fmt.Errorf("runner[%v] %v convertBool for %v (%v) error %v, this key will be ignored", r.meta.RunnerName, r.Name(), columns[i], scanArgs[i], serr)
-				log.Error(serr)
-				r.sendError(serr)
-			} else {
-				data[columns[i]] = val
-				bytes = 4
-			}
-		case "date":
-			val, serr := convertDate(scanArgs[i])
-			if serr != nil {
-				serr = fmt.Errorf("runner[%v] %v convertDate for %v (%v) error %v, this key will be ignored", r.meta.RunnerName, r.Name(), columns[i], scanArgs[i], serr)
-				log.Error(serr)
-				r.sendError(serr)
-			} else {
-				data[columns[i]] = val
-				bytes = 20
-			}
-		default:
-			dealed := false
-			if !nochiced[i] {
-				dealed = true
-				switch d := scanArgs[i].(type) {
-				case *string:
-					data[columns[i]] = *d
-					bytes = int64(len(*d))
-				case *[]byte:
-					data[columns[i]] = string(*d)
-					bytes = int64(len(*d))
-				case *bool:
-					data[columns[i]] = *d
-					bytes = 1
-				case int64:
-					data[columns[i]] = d
-					bytes = 8
-				case *int64:
-					data[columns[i]] = *d
-					bytes = 8
-				case float64:
-					data[columns[i]] = d
-					bytes = 8
-				case *float64:
-					data[columns[i]] = *d
-					bytes = 8
-				case uint64:
-					data[columns[i]] = d
-					bytes = 8
-				case *uint64:
-					data[columns[i]] = *d
-					bytes = 8
-				case *interface{}:
-					dealed = false
-				default:
-					dealed = false
-				}
-			}
-			if !dealed {
-				val, serr := convertString(scanArgs[i])
-				if serr != nil {
-					serr = fmt.Errorf("runner[%v] %v convertString for %v (%v) error %v, this key will be ignored", r.meta.RunnerName, r.Name(), columns[i], scanArgs[i], serr)
-					log.Error(serr)
-					r.sendError(serr)
-				} else {
-					data[columns[i]] = val
-					bytes = int64(len(val))
-				}
-			}
 		}
-
-		totalBytes += bytes
 	}
-	return data, totalBytes
+	return bytes, err
 }
