+++ reader/sql/magic.go
@@ -9,7 +9,7 @@ import (
 )
 
 // 仅支持 YYYY, YY, MM, DD, hh, mm, ss，不支持 M, D, h, m, s
-func convertMagicIndex(magic string, now time.Time) (ret string, index int) {
+func ConvertMagicIndex(magic string, now time.Time) (Ret string, index int) {
 	switch magic {
 	case "YYYY":
 		return fmt.Sprintf("%d", now.Year()), YEAR
@@ -35,65 +35,65 @@ func convertMagicIndex(magic string, now time.Time) (ret string, index int) {
 }
 
 type MagicRes struct {
-	timeStart   []int  // 按照 YY,MM,DD,hh,mm,ss 顺序记录时间位置
-	timeEnd     []int  // 按照 YY,MM,DD,hh,mm,ss 顺序记录时间长度
-	remainIndex []int  // 按顺序记录非时间字符串开始结束位置，去除 *
-	ret         string // 渲染结果，包含 *
+	TimeStart   []int  // 按照 YY,MM,DD,hh,mm,ss 顺序记录时间位置
+	TimeEnd     []int  // 按照 YY,MM,DD,hh,mm,ss 顺序记录时间长度
+	RemainIndex []int  // 按顺序记录非时间字符串开始结束位置，去除 *
+	Ret         string // 渲染结果，包含 *
 }
 
 // 渲染魔法变量
-func goMagicIndex(rawSql string, now time.Time) (MagicRes, error) {
+func GoMagicIndex(rawSql string, now time.Time) (MagicRes, error) {
 	sps := strings.Split(rawSql, "@(") //@()，对于每个分片找右括号
 	var magicRes = MagicRes{
-		timeStart:   []int{-1, -1, -1, -1, -1, -1},
-		timeEnd:     make([]int, 6),
-		remainIndex: []int{0},
-		ret:         sps[0],
+		TimeStart:   []int{-1, -1, -1, -1, -1, -1},
+		TimeEnd:     make([]int, 6),
+		RemainIndex: []int{0},
+		Ret:         sps[0],
 	}
-	recordIndex := len(magicRes.ret)
+	recordIndex := len(magicRes.Ret)
 
 	// 没有魔法变量的情况，例如 mytest*
 	if len(sps) < 2 {
-		magicRes.remainIndex = append(magicRes.remainIndex, removeWildcards(magicRes.ret, recordIndex))
+		magicRes.RemainIndex = append(magicRes.RemainIndex, RemoveWildcards(magicRes.Ret, recordIndex))
 		return magicRes, nil
 	}
 
-	magicRes.remainIndex = append(magicRes.remainIndex, recordIndex)
+	magicRes.RemainIndex = append(magicRes.RemainIndex, recordIndex)
 	for idx := 1; idx < len(sps); idx++ {
 		idxr := strings.Index(sps[idx], ")")
 		if idxr == -1 {
-			magicRes.ret = rawSql
+			magicRes.Ret = rawSql
 			return magicRes, nil
 		}
 		spsStr := sps[idx][0:idxr]
 		if len(spsStr) < 2 {
-			magicRes.ret = rawSql
+			magicRes.Ret = rawSql
 			return magicRes, fmt.Errorf(SupportReminder)
 		}
-		res, index := convertMagicIndex(sps[idx][0:idxr], now)
+		res, index := ConvertMagicIndex(sps[idx][0:idxr], now)
 		if index == -1 {
-			magicRes.ret = rawSql
+			magicRes.Ret = rawSql
 			return magicRes, fmt.Errorf(SupportReminder)
 		}
 
 		// 记录日期起始位置
-		magicRes.timeStart[index] = recordIndex
-		magicRes.ret += res
-		recordIndex = len(magicRes.ret)
+		magicRes.TimeStart[index] = recordIndex
+		magicRes.Ret += res
+		recordIndex = len(magicRes.Ret)
 
 		// 记录日期长度
-		magicRes.timeEnd[index] = recordIndex
+		magicRes.TimeEnd[index] = recordIndex
 
 		if idxr+1 < len(sps[idx]) {
 			spsRemain := sps[idx][idxr+1:]
-			magicRes.ret += spsRemain
+			magicRes.Ret += spsRemain
 			if spsRemain == Wildcards {
-				recordIndex = len(magicRes.ret)
+				recordIndex = len(magicRes.Ret)
 				continue
 			}
-			magicRes.remainIndex = append(magicRes.remainIndex, recordIndex)
-			magicRes.remainIndex = append(magicRes.remainIndex, removeWildcards(spsRemain, len(magicRes.ret)))
-			recordIndex = len(magicRes.ret)
+			magicRes.RemainIndex = append(magicRes.RemainIndex, recordIndex)
+			magicRes.RemainIndex = append(magicRes.RemainIndex, RemoveWildcards(spsRemain, len(magicRes.Ret)))
+			recordIndex = len(magicRes.Ret)
 		}
 	}
 
@@ -101,14 +101,14 @@ func goMagicIndex(rawSql string, now time.Time) (MagicRes, error) {
 }
 
 // 若包含通配符，字段长度相应 - 1
-func removeWildcards(checkWildcards string, length int) int {
+func RemoveWildcards(checkWildcards string, length int) int {
 	if strings.Contains(checkWildcards, Wildcards) {
 		return length - 1
 	}
 	return length
 }
 
-func checkMagic(rawSql string) (valid bool) {
+func CheckMagic(rawSql string) (valid bool) {
 	sps := strings.Split(rawSql, "@(") //@()，对于每个分片找右括号
 	now := time.Now()
 
@@ -123,7 +123,7 @@ func checkMagic(rawSql string) (valid bool) {
 			return false
 		}
 
-		_, index := convertMagicIndex(sps[idx][0:idxr], now)
+		_, index := ConvertMagicIndex(sps[idx][0:idxr], now)
 		if index == -1 {
 			log.Errorf(SupportReminder)
 			return false
