--- reader/sql/sql_test.go
@@ -1,2272 +1,0 @@
-package sql
-
-import (
-	"database/sql"
-	"fmt"
-	"io/ioutil"
-	"os"
-	"path"
-	"reflect"
-	"strconv"
-	"strings"
-	"sync"
-	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
-
-	"github.com/qiniu/log"
-	"github.com/qiniu/logkit/conf"
-	"github.com/qiniu/logkit/reader"
-	. "github.com/qiniu/logkit/reader/config"
-	"github.com/qiniu/logkit/reader/sql/datagen"
-	. "github.com/qiniu/logkit/reader/test"
-	. "github.com/qiniu/logkit/utils/models"
-)
-
-var readRecords = DBRecords{
-	"db1": TableRecords{
-		Table: map[string]TableInfo{
-			"db1_tb1": {size: -1, offset: -1},
-			"db1_tb2": {size: -1, offset: -1},
-			"db1_tb3": {size: -1, offset: -1},
-		},
-	},
-	"db2": TableRecords{
-		Table: map[string]TableInfo{
-			"db2_tb1": {size: -1, offset: -1},
-			"db2_tb2": {size: -1, offset: -1},
-			"db2_tb3": {size: -1, offset: -1},
-			"db2_tb4": {size: -1, offset: -1},
-			"db2_tb5": {size: -1, offset: -1},
-		},
-	},
-	"db3": TableRecords{
-		Table: map[string]TableInfo{
-			"db3_tb1": {size: -1, offset: -1},
-			"db3_tb2": {size: -1, offset: -1},
-		},
-	},
-}
-
-func TestConvertMagicIndex(t *testing.T) {
-	now1, _ := time.Parse(time.RFC3339, "2017-04-01T06:06:09+08:00")
-	now2, _ := time.Parse(time.RFC3339, "2017-11-11T16:16:29+08:00")
-	tests := []struct {
-		data     string
-		exp1     string
-		exp2     string
-		expIndex int
-	}{
-		{
-			data:     "YYYY",
-			exp1:     "2017",
-			exp2:     "2017",
-			expIndex: YEAR,
-		},
-		{
-			data:     "YY",
-			exp1:     "17",
-			exp2:     "17",
-			expIndex: YEAR,
-		},
-		{
-			data:     "MM",
-			exp1:     "04",
-			exp2:     "11",
-			expIndex: MONTH,
-		},
-		{
-			data:     "DD",
-			exp1:     "01",
-			exp2:     "11",
-			expIndex: DAY,
-		},
-		{
-			data:     "hh",
-			exp1:     "06",
-			exp2:     "16",
-			expIndex: HOUR,
-		},
-		{
-			data:     "mm",
-			exp1:     "06",
-			exp2:     "16",
-			expIndex: MINUTE,
-		},
-		{
-			data:     "m",
-			exp1:     "",
-			exp2:     "",
-			expIndex: -1,
-		},
-		{
-			data:     "ss",
-			exp1:     "09",
-			exp2:     "29",
-			expIndex: SECOND,
-		},
-		{
-			data:     "s",
-			exp1:     "",
-			exp2:     "",
-			expIndex: -1,
-		},
-	}
-	for _, ti := range tests {
-		got, gotIndex := convertMagicIndex(ti.data, now1)
-		assert.Equal(t, ti.exp1, got)
-		assert.Equal(t, ti.expIndex, gotIndex)
-		got, gotIndex = convertMagicIndex(ti.data, now2)
-		assert.Equal(t, ti.exp2, got)
-		assert.Equal(t, ti.expIndex, gotIndex)
-	}
-}
-
-func TestGoMagicIndex(t *testing.T) {
-	now, _ := time.Parse(time.RFC3339, "2017-02-01T16:06:19+08:00")
-	tests := []struct {
-		data          string
-		expRet        string
-		expStartIndex []int
-		expEndIndex   []int
-		expTimeIndex  []int
-	}{
-		{
-			data:          "x@(MM)abc@(DD)def",
-			expRet:        "x02abc01def",
-			expStartIndex: []int{-1, 1, 6, -1, -1, -1},
-			expEndIndex:   []int{0, 3, 8, 0, 0, 0},
-			expTimeIndex:  []int{0, 1, 3, 6, 8, 11},
-		},
-		{
-			data:          "x@(MM)abc@(DD)def*",
-			expRet:        "x02abc01def*",
-			expStartIndex: []int{-1, 1, 6, -1, -1, -1},
-			expEndIndex:   []int{0, 3, 8, 0, 0, 0},
-			expTimeIndex:  []int{0, 1, 3, 6, 8, 11},
-		},
-		{
-			data:          "x@(MM)abc@(DD)",
-			expRet:        "x02abc01",
-			expStartIndex: []int{-1, 1, 6, -1, -1, -1},
-			expEndIndex:   []int{0, 3, 8, 0, 0, 0},
-			expTimeIndex:  []int{0, 1, 3, 6},
-		},
-		{
-			data:          "x@(MM)@(DD)",
-			expRet:        "x0201",
-			expStartIndex: []int{-1, 1, 3, -1, -1, -1},
-			expEndIndex:   []int{0, 3, 5, 0, 0, 0},
-			expTimeIndex:  []int{0, 1},
-		},
-		{
-			data:          "x@(DD)@(MM)*",
-			expRet:        "x0102*",
-			expStartIndex: []int{-1, 3, 1, -1, -1, -1},
-			expEndIndex:   []int{0, 5, 3, 0, 0, 0},
-			expTimeIndex:  []int{0, 1},
-		},
-		{
-			data:          "@(YY)",
-			expRet:        "17",
-			expStartIndex: []int{0, -1, -1, -1, -1, -1},
-			expEndIndex:   []int{2, 0, 0, 0, 0, 0},
-			expTimeIndex:  []int{0, 0},
-		},
-		{
-			data:          "abcd@(YYYY)@(MM)efg*",
-			expRet:        "abcd201702efg*",
-			expStartIndex: []int{4, 8, -1, -1, -1, -1},
-			expEndIndex:   []int{8, 10, 0, 0, 0, 0},
-			expTimeIndex:  []int{0, 4, 10, 13},
-		},
-		{
-			data:          "abcd@(YYYY)@(MM)@(DD)@(hh)@(mm)@(ss)*",
-			expRet:        "abcd20170201160619*",
-			expStartIndex: []int{4, 8, 10, 12, 14, 16},
-			expEndIndex:   []int{8, 10, 12, 14, 16, 18},
-			expTimeIndex:  []int{0, 4},
-		},
-		{
-			data:          "hhhhh",
-			expRet:        "hhhhh",
-			expStartIndex: []int{-1, -1, -1, -1, -1, -1},
-			expEndIndex:   []int{0, 0, 0, 0, 0, 0},
-			expTimeIndex:  []int{0, 5},
-		},
-	}
-	for _, ti := range tests {
-		magicRes, err := goMagicIndex(ti.data, now)
-		assert.NoError(t, err)
-		assert.EqualValues(t, ti.expRet, magicRes.ret)
-		assert.EqualValues(t, ti.expStartIndex, magicRes.timeStart)
-		assert.EqualValues(t, ti.expEndIndex, magicRes.timeEnd)
-		assert.EqualValues(t, ti.expTimeIndex, magicRes.remainIndex)
-	}
-
-	errData := "x@(M)@(DD)"
-	expRet := "x@(M)@(DD)"
-	magicRes, err := goMagicIndex(errData, now)
-	assert.Error(t, err)
-	assert.EqualValues(t, expRet, magicRes.ret)
-}
-
-func Test_getRemainStr(t *testing.T) {
-	tests := []struct {
-		origin       string
-		timeIndex    []int
-		expectRemain string
-	}{
-		{
-			origin:       "x02abc01def",
-			timeIndex:    []int{0, 1, 3, 6, 8, 11},
-			expectRemain: "xabcdef",
-		},
-		{
-			origin:       "x02abc01def*",
-			timeIndex:    []int{0, 1, 3, 6, 8, 11},
-			expectRemain: "xabcdef",
-		},
-		{
-			origin:       "x02abc01",
-			timeIndex:    []int{0, 1, 3, 6},
-			expectRemain: "xabc",
-		},
-		{
-			origin:       "x0201",
-			timeIndex:    []int{0, 1},
-			expectRemain: "x",
-		},
-		{
-			origin:       "x0102*",
-			timeIndex:    []int{0, 1},
-			expectRemain: "x",
-		},
-		{
-			origin:       "17",
-			timeIndex:    []int{0, 0},
-			expectRemain: "",
-		},
-		{
-			origin:       "abcd201702efg*",
-			timeIndex:    []int{0, 4, 10, 13},
-			expectRemain: "abcdefg",
-		},
-		{
-			origin:       "abcd20170201160619*",
-			timeIndex:    []int{0, 4},
-			expectRemain: "abcd",
-		},
-		{
-			origin:       "hhhhh",
-			timeIndex:    []int{0, 5},
-			expectRemain: "hhhhh",
-		},
-	}
-	for _, ti := range tests {
-		remain := getRemainStr(ti.origin, ti.timeIndex)
-		assert.Equal(t, ti.expectRemain, remain)
-	}
-}
-
-func Test_matchRemainStr(t *testing.T) {
-	tests := []struct {
-		origin    string
-		match     string
-		matchData string
-		timeIndex []int
-		expectRes bool
-	}{
-		{
-			origin:    "x02abc01def",
-			match:     "xabcdef",
-			matchData: "x02abc01def",
-			timeIndex: []int{0, 1, 3, 6, 8, 11},
-			expectRes: true,
-		},
-		{
-			origin:    "x02abc01defdef",
-			match:     "xabcdef",
-			matchData: "x02abc01def*",
-			timeIndex: []int{0, 1, 3, 6, 8, 12},
-			expectRes: true,
-		},
-		{
-			origin:    "x02abc01",
-			match:     "xabc",
-			matchData: "x02abc01",
-			timeIndex: []int{0, 1, 3, 6},
-			expectRes: true,
-		},
-		{
-			origin:    "x0201",
-			match:     "x",
-			matchData: "x0201",
-			timeIndex: []int{0, 1},
-			expectRes: true,
-		},
-		{
-			origin:    "x0102*",
-			match:     "x*",
-			matchData: "x0102*",
-			timeIndex: []int{0, 1, 5, 6},
-			expectRes: true,
-		},
-		{
-			origin:    "17",
-			match:     "",
-			matchData: "17",
-			timeIndex: []int{0, 0},
-			expectRes: true,
-		},
-		{
-			origin:    "abcd201702efg*",
-			match:     "xabcdef",
-			matchData: "abcd201702efg*",
-			timeIndex: []int{0, 4, 10, 14},
-			expectRes: false,
-		},
-		{
-			origin:    "abcd20170201160619*",
-			match:     "xabcdef",
-			matchData: "abcd20170201160619*",
-			timeIndex: []int{0, 4, 18, 19},
-			expectRes: false,
-		},
-		{
-			origin:    "abcd20170201160619ef",
-			match:     "abcd",
-			matchData: "abcd20170201160619",
-			timeIndex: []int{0, 4},
-			expectRes: false,
-		},
-		{
-			origin:    "hhhhh",
-			match:     "hhhhh",
-			matchData: "hhhhh",
-			timeIndex: []int{0, 5},
-			expectRes: true,
-		},
-		{
-			origin:    "hhhh",
-			match:     "hhhhh",
-			matchData: "hhhhh",
-			timeIndex: []int{0, 5},
-			expectRes: false,
-		},
-	}
-	for _, ti := range tests {
-		remain := compareRemainStr(ti.origin, ti.match, ti.matchData, ti.timeIndex)
-		assert.Equal(t, ti.expectRes, remain)
-	}
-}
-
-func Test_checkMagic(t *testing.T) {
-	tests := []struct {
-		data string
-		exp  bool
-	}{
-		{
-			data: "x@(DD)@(MM)*",
-			exp:  true,
-		},
-		{
-			data: "@(YY)",
-			exp:  true,
-		},
-		{
-			data: "abcd@(YYYY)@(M)efg*",
-			exp:  false,
-		},
-		{
-			data: "abcd@(YYYY)@(MM)@(DD)@(hh)@(mm)@(ss)efg*",
-			exp:  true,
-		},
-		{
-			data: "hhhhh",
-			exp:  true,
-		},
-	}
-	for _, ti := range tests {
-		got := checkMagic(ti.data)
-		assert.EqualValues(t, ti.exp, got)
-	}
-}
-
-func TestSQLReader(t *testing.T) {
-	meta, err := getMeta(MetaDir)
-	assert.NoError(t, err)
-	defer os.RemoveAll(MetaDir)
-	database := "TestSQLReaderdatabase"
-	mr := &Reader{
-		rawDatabase: database,
-		database:    database,
-		rawSQLs:     "select * from mysql123  ;select * from mysql345;",
-		syncSQLs:    []string{"select * from mysql123", "select * from mysql345"},
-		readBatch:   100,
-		meta:        meta,
-		offsetKey:   "id",
-		offsets:     []int64{123, 456},
-		dbtype:      "mysql",
-	}
-	assert.Equal(t, mr.dbtype+"_"+database, mr.Source())
-
-	// 测试meta备份和恢复
-	mr.SyncMeta()
-	gotoffsets, gotsqls, omit := restoreMeta(meta, mr.rawSQLs, 0)
-	assert.EqualValues(t, mr.offsets, gotoffsets, "got offsets error")
-	assert.EqualValues(t, mr.syncSQLs, gotsqls, "got sqls error")
-	assert.EqualValues(t, false, omit)
-	assert.EqualValues(t, "MYSQL_Reader:"+mr.database+"_"+Hash(mr.rawSQLs), mr.Name())
-
-	// 测试更新Offset
-	expoffsets := []int64{123, 0, 0}
-	testsqls := []string{"select * from mysql123", "select * from mysql789", "select x from xx"}
-	mr.updateOffsets(testsqls)
-	assert.EqualValues(t, expoffsets, mr.offsets)
-	mr.syncSQLs = testsqls
-
-	//测试getSQL
-	gotSQL, err := mr.getSQL(2, mr.syncSQLs[2])
-	assert.NoError(t, err)
-	assert.EqualValues(t, testsqls[2]+" WHERE id >= 0 AND id < 100;", gotSQL)
-	mr.offsetKey = ""
-	gotSQL, err = mr.getSQL(0, mr.syncSQLs[0])
-	assert.NoError(t, err)
-	assert.EqualValues(t, testsqls[0], gotSQL)
-
-	assert.Equal(t, StatsInfo{}, mr.Status())
-}
-
-func TestUpdateSql(t *testing.T) {
-	rawsqls := "select * from mysql123  ;select * from mysql345;"
-	syncSQLs := []string{"select * from mysql123", "select * from mysql345"}
-	got := updateSqls(rawsqls, time.Now())
-	assert.EqualValues(t, syncSQLs, got)
-}
-
-func TestReflectTime(t *testing.T) {
-	fmt.Println(reflect.TypeOf(int64(0)))
-	fmt.Println(reflect.TypeOf(int32(0)))
-	fmt.Println(reflect.TypeOf(int16(0)))
-	fmt.Println(reflect.TypeOf(""))
-	fmt.Println(reflect.TypeOf(false))
-	fmt.Println(reflect.TypeOf(time.Time{}))
-	fmt.Println(reflect.TypeOf([]byte{}))
-	fmt.Println(reflect.TypeOf(new(interface{})).Elem())
-}
-
-func Test_getDefaultSql(t *testing.T) {
-	database := "my_database"
-	r1 := Reader{
-		dbtype: ModeMySQL,
-	}
-	actualSql, err := r1.getDefaultSql(database)
-	assert.NoError(t, err)
-	expectSql := strings.Replace(DefaultMySQLTable, "DATABASE_NAME", database, -1)
-	assert.Equal(t, actualSql, expectSql)
-
-	r2 := Reader{
-		dbtype:   ModeMSSQL,
-		dbSchema: "dbo",
-	}
-	actualSql, err = r2.getDefaultSql(database)
-	assert.NoError(t, err)
-	sql := strings.Replace(DefaultMSSQLTable, "DATABASE_NAME", database, -1)
-	expectSql = strings.Replace(sql, "SCHEMA_NAME", r2.dbSchema, -1)
-	assert.Equal(t, actualSql, expectSql)
-
-	r3 := Reader{
-		dbtype:   ModePostgreSQL,
-		dbSchema: "public",
-	}
-	actualSql, err = r3.getDefaultSql(database)
-	assert.NoError(t, err)
-	sql = strings.Replace(DefaultPGSQLTable, "DATABASE_NAME", database, -1)
-	expectSql = strings.Replace(sql, "SCHEMA_NAME", r3.dbSchema, -1)
-	assert.Equal(t, actualSql, expectSql)
-}
-
-func Test_restoreMeta(t *testing.T) {
-	meta, err := reader.NewMeta(MetaDir, MetaDir, "mysql", "logpath", "", 7)
-	if err != nil {
-		t.Error(err)
-	}
-	defer os.RemoveAll(MetaDir)
-
-	file, offset, err := meta.ReadOffset()
-	if err == nil {
-		t.Error("offset must be nil")
-	}
-	encodeSQLs := make([]string, 0)
-	syncSQLs := []string{"SELECT * FROM A", "SELECT * FROM B", "SELECT * FROM C"}
-	offsets := []int64{1, 2, 3}
-	for _, sql := range syncSQLs {
-		encodeSQLs = append(encodeSQLs, strings.Replace(sql, " ", "@", -1))
-	}
-	for _, offset := range offsets {
-		encodeSQLs = append(encodeSQLs, strconv.FormatInt(offset, 10))
-	}
-	all := strings.Join(encodeSQLs, sqlOffsetConnector)
-	err = meta.WriteOffset(all, int64(len(syncSQLs)))
-	if err != nil {
-		t.Error(err)
-	}
-	file, offset, err = meta.ReadOffset()
-	if file != all {
-		t.Error("sql meta offset should be " + all)
-	}
-	if offset != 3 {
-		t.Error("file offset should be 3")
-	}
-
-	var mgld time.Duration
-	rawSqls := "SELECT * FROM A; SELECT * FROM B; SELECT * FROM C;"
-	actualOffsets, actualSqls, omitMeta := restoreMeta(meta, rawSqls, mgld)
-	assert.EqualValues(t, false, omitMeta)
-	assert.EqualValues(t, offsets, actualOffsets)
-	assert.EqualValues(t, syncSQLs, actualSqls)
-
-	rawSqls = "SELECT * FROM A; SELECT * FROM B; SELECT * FROM D;"
-	actualOffsets, actualSqls, omitMeta = restoreMeta(meta, rawSqls, mgld)
-	assert.EqualValues(t, false, omitMeta)
-	assert.EqualValues(t, offsets, actualOffsets)
-	assert.EqualValues(t, actualSqls, actualSqls)
-
-	rawSqls = "SELECT * FROM A; SELECT * FROM B;"
-	actualOffsets, actualSqls, omitMeta = restoreMeta(meta, rawSqls, mgld)
-	assert.EqualValues(t, true, omitMeta)
-}
-
-func Test_validTime(t *testing.T) {
-	tests := []struct {
-		data       string
-		match      string
-		startIndex []int
-		endIndex   []int
-		expRes     bool
-	}{
-		{
-			data:       "x02abc01",
-			match:      "x02abc01",
-			startIndex: []int{-1, 1, 6, -1, -1, -1},
-			endIndex:   []int{0, 3, 8, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "hhhhh",
-			match:      "hhhhh",
-			startIndex: []int{-1, -1, -1, -1, -1, -1},
-			endIndex:   []int{0, 0, 0, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "x01abc31def",
-			match:      "x02abc01def",
-			startIndex: []int{-1, 1, 6, -1, -1, -1},
-			endIndex:   []int{0, 3, 8, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "x01abc31def*",
-			match:      "x02abc01def*",
-			startIndex: []int{-1, 1, 6, -1, -1, -1},
-			endIndex:   []int{0, 3, 8, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "x0201",
-			match:      "x0201",
-			startIndex: []int{-1, 1, 3, -1, -1, -1},
-			endIndex:   []int{0, 3, 5, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "x0102abc",
-			match:      "x0102*",
-			startIndex: []int{-1, 3, 1, -1, -1, -1},
-			endIndex:   []int{0, 5, 3, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "x0102",
-			match:      "x0102*",
-			startIndex: []int{-1, 3, 1, -1, -1, -1},
-			endIndex:   []int{0, 5, 3, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "17",
-			match:      "17",
-			startIndex: []int{0, -1, -1, -1, -1, -1},
-			endIndex:   []int{2, 0, 0, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "abcd201703efg*",
-			match:      "abcd201702efg*",
-			startIndex: []int{4, 8, -1, -1, -1, -1},
-			endIndex:   []int{8, 10, 0, 0, 0, 0},
-			expRes:     false,
-		},
-		{
-			data:       "abcd20170201160618*",
-			match:      "abcd20170201160619*",
-			startIndex: []int{4, 8, 10, 12, 14, 16},
-			endIndex:   []int{8, 10, 12, 14, 16, 18},
-			expRes:     true,
-		},
-	}
-	for _, ti := range tests {
-		valid := compareTime(ti.data, ti.match, ti.startIndex, ti.endIndex, true)
-		assert.EqualValues(t, ti.expRes, valid)
-	}
-
-	tests2 := []struct {
-		data       string
-		match      string
-		startIndex []int
-		endIndex   []int
-		expRes     bool
-	}{
-		{
-			data:       "x02abc01",
-			match:      "x02abc01",
-			startIndex: []int{-1, 1, 6, -1, -1, -1},
-			endIndex:   []int{0, 3, 8, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "hhhhh",
-			match:      "hhhhh",
-			startIndex: []int{-1, -1, -1, -1, -1, -1},
-			endIndex:   []int{0, 0, 0, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "x01abc31def",
-			match:      "x02abc01def",
-			startIndex: []int{-1, 1, 6, -1, -1, -1},
-			endIndex:   []int{0, 3, 8, 0, 0, 0},
-			expRes:     false,
-		},
-		{
-			data:       "x01abc31def*",
-			match:      "x02abc01def*",
-			startIndex: []int{-1, 1, 6, -1, -1, -1},
-			endIndex:   []int{0, 3, 8, 0, 0, 0},
-			expRes:     false,
-		},
-		{
-			data:       "x0201",
-			match:      "x0201",
-			startIndex: []int{-1, 1, 3, -1, -1, -1},
-			endIndex:   []int{0, 3, 5, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "x0102abc",
-			match:      "x0102*",
-			startIndex: []int{-1, 3, 1, -1, -1, -1},
-			endIndex:   []int{0, 5, 3, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "x0102",
-			match:      "x0102*",
-			startIndex: []int{-1, 3, 1, -1, -1, -1},
-			endIndex:   []int{0, 5, 3, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "17",
-			match:      "17",
-			startIndex: []int{0, -1, -1, -1, -1, -1},
-			endIndex:   []int{2, 0, 0, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "abcd201703efg*",
-			match:      "abcd201702efg*",
-			startIndex: []int{4, 8, -1, -1, -1, -1},
-			endIndex:   []int{8, 10, 0, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "abcd20170201160618*",
-			match:      "abcd20170201160619*",
-			startIndex: []int{4, 8, 10, 12, 14, 16},
-			endIndex:   []int{8, 10, 12, 14, 16, 18},
-			expRes:     false,
-		},
-	}
-	for _, ti := range tests2 {
-		valid := compareTime(ti.data, ti.match, ti.startIndex, ti.endIndex, false)
-		assert.EqualValues(t, ti.expRes, valid)
-	}
-}
-
-func Test_equalTime(t *testing.T) {
-	tests := []struct {
-		data       string
-		match      string
-		startIndex []int
-		endIndex   []int
-		expRes     bool
-	}{
-		{
-			data:       "x02abc01",
-			match:      "x02abc01",
-			startIndex: []int{-1, 1, 6, -1, -1, -1},
-			endIndex:   []int{0, 3, 8, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "hhhhh",
-			match:      "hhhhh",
-			startIndex: []int{-1, -1, -1, -1, -1, -1},
-			endIndex:   []int{0, 0, 0, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "x01abc31def",
-			match:      "x02abc01def",
-			startIndex: []int{-1, 1, 6, -1, -1, -1},
-			endIndex:   []int{0, 3, 8, 0, 0, 0},
-			expRes:     false,
-		},
-		{
-			data:       "x0201",
-			match:      "x0201",
-			startIndex: []int{-1, 1, 3, -1, -1, -1},
-			endIndex:   []int{0, 3, 5, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "17",
-			match:      "17",
-			startIndex: []int{0, -1, -1, -1, -1, -1},
-			endIndex:   []int{2, 0, 0, 0, 0, 0},
-			expRes:     true,
-		},
-		{
-			data:       "abcd201703efg*",
-			match:      "abcd201702efg*",
-			startIndex: []int{4, 8, -1, -1, -1, -1},
-			endIndex:   []int{8, 10, 0, 0, 0, 0},
-			expRes:     false,
-		},
-		{
-			data:       "abcd20170201160618*",
-			match:      "abcd20170201160619*",
-			startIndex: []int{4, 8, 10, 12, 14, 16},
-			endIndex:   []int{8, 10, 12, 14, 16, 18},
-			expRes:     false,
-		},
-	}
-	for _, ti := range tests {
-		valid := equalTime(ti.data, ti.match, ti.startIndex, ti.endIndex)
-		assert.EqualValues(t, ti.expRes, valid)
-	}
-}
-
-func Test_isMatchData(t *testing.T) {
-	meta, err := getMeta(MetaDir)
-	assert.NoError(t, err)
-	defer os.RemoveAll(MetaDir)
-	mr := &Reader{
-		readBatch:         100,
-		meta:              meta,
-		dbtype:            "mysql",
-		historyAll:        false,
-		isLoop:            true,
-		cronSchedule:      false,
-		omitDoneDBRecords: true,
-	}
-
-	tests := []struct {
-		data           string
-		magicRemainStr string
-		magicRet       string
-		expRet         bool
-		expTimeStart   []int
-		exptimeEnd     []int
-		expRemainIndex []int
-	}{
-		{
-			data:           "x02abc01def",
-			magicRemainStr: "xabcdef",
-			magicRet:       "x02abc01def",
-			expRet:         true,
-			expTimeStart:   []int{-1, 1, 6, -1, -1, -1},
-			exptimeEnd:     []int{0, 3, 8, 0, 0, 0},
-			expRemainIndex: []int{0, 1, 3, 6, 8, 11},
-		},
-		{
-			data:           "x01abc01def",
-			magicRemainStr: "xabcdef",
-			magicRet:       "x02abc01def",
-			expRet:         true,
-			expTimeStart:   []int{-1, 1, 6, -1, -1, -1},
-			exptimeEnd:     []int{0, 3, 8, 0, 0, 0},
-			expRemainIndex: []int{0, 1, 3, 6, 8, 11},
-		},
-		{
-			data:           "x02abc01defdef",
-			magicRemainStr: "xabcdef",
-			magicRet:       "x02abc01def*",
-			expRet:         true,
-			expTimeStart:   []int{-1, 1, 6, -1, -1, -1},
-			exptimeEnd:     []int{0, 3, 8, 0, 0, 0},
-			expRemainIndex: []int{0, 1, 3, 6, 8, 11},
-		},
-		{
-			data:           "abc",
-			magicRemainStr: "xabc",
-			magicRet:       "x02abc01",
-			expRet:         false,
-			expTimeStart:   []int{-1, 1, 6, -1, -1, -1},
-			exptimeEnd:     []int{0, 3, 8, 0, 0, 0},
-			expRemainIndex: []int{0, 1, 3, 6},
-		},
-		{
-			data:           "x0201",
-			magicRemainStr: "x",
-			magicRet:       "x0201*",
-			expRet:         true,
-			expTimeStart:   []int{-1, 1, 3, -1, -1, -1},
-			exptimeEnd:     []int{0, 3, 5, 0, 0, 0},
-			expRemainIndex: []int{0, 1},
-		},
-		{
-			data:           "x0102xxxxx",
-			magicRemainStr: "x",
-			magicRet:       "x0102*",
-			expRet:         true,
-			expTimeStart:   []int{-1, 3, 1, -1, -1, -1},
-			exptimeEnd:     []int{0, 5, 3, 0, 0, 0},
-			expRemainIndex: []int{0, 1},
-		},
-		{
-			data:           "17",
-			magicRemainStr: "",
-			magicRet:       "17",
-			expRet:         true,
-			expTimeStart:   []int{0, -1, -1, -1, -1, -1},
-			exptimeEnd:     []int{2, 0, 0, 0, 0, 0},
-			expRemainIndex: []int{0, 0},
-		},
-		{
-			data:           "abcd201702efg",
-			magicRemainStr: "abcdefg",
-			magicRet:       "abcd201702efg*",
-			expRet:         true,
-			expTimeStart:   []int{4, 8, -1, -1, -1, -1},
-			exptimeEnd:     []int{8, 10, 0, 0, 0, 0},
-			expRemainIndex: []int{0, 4, 10, 13},
-		},
-		{
-			data:           "abcd20170201160619abc",
-			magicRemainStr: "abcd",
-			magicRet:       "abcd20170201160619",
-			expRet:         false,
-			expTimeStart:   []int{4, 8, 10, 12, 14, 16},
-			exptimeEnd:     []int{8, 10, 12, 14, 16, 18},
-			expRemainIndex: []int{0, 4},
-		},
-		{
-			data:           "hhhhh",
-			magicRemainStr: "hhhhh",
-			magicRet:       "hhhhh",
-			expRet:         true,
-			expTimeStart:   []int{-1, -1, -1, -1, -1, -1},
-			exptimeEnd:     []int{0, 0, 0, 0, 0, 0},
-			expRemainIndex: []int{0, 5},
-		},
-	}
-	for _, ti := range tests {
-		ret := mr.compareData(DATABASE, "", ti.data, ti.magicRemainStr, &MagicRes{
-			ret:         ti.magicRet,
-			remainIndex: ti.expRemainIndex,
-			timeStart:   ti.expTimeStart,
-			timeEnd:     ti.exptimeEnd,
-		})
-		assert.EqualValues(t, ti.expRet, ret)
-	}
-}
-
-func Test_getCheckAll(t *testing.T) {
-	mr := &Reader{
-		database:    "Test_getCheckHistory",
-		dbtype:      "mysql",
-		historyAll:  true,
-		rawTable:    "*",
-		rawDatabase: "*",
-	}
-
-	tests := []struct {
-		queryType int
-		expRes    bool
-	}{
-		{
-			queryType: TABLE,
-			expRes:    true,
-		},
-		{
-			queryType: COUNT,
-			expRes:    true,
-		},
-		{
-			queryType: DATABASE,
-			expRes:    true,
-		},
-	}
-
-	for _, test := range tests {
-		checkHistory, err := mr.getAll(test.queryType)
-		assert.NoError(t, err)
-		assert.EqualValues(t, test.expRes, checkHistory)
-	}
-}
-func Test_getWrappedTableName(t *testing.T) {
-	r1 := Reader{
-		dbtype: ModeMySQL,
-	}
-	tname, err := r1.getWrappedTableName("my_table")
-	expRes := "`my_table`"
-	assert.NoError(t, err)
-	assert.EqualValues(t, expRes, tname)
-
-	r2 := Reader{
-		dbtype:   ModeMSSQL,
-		dbSchema: "dbo",
-	}
-	tname, err = r2.getWrappedTableName("my_table")
-	expRes = "\"dbo\".\"my_table\""
-	assert.NoError(t, err)
-	assert.EqualValues(t, expRes, tname)
-
-	r3 := Reader{
-		dbtype:   ModePostgreSQL,
-		dbSchema: "public",
-	}
-	tname, err = r3.getWrappedTableName("my_table")
-	expRes = "\"public\".\"my_table\""
-	assert.NoError(t, err)
-	assert.EqualValues(t, expRes, tname)
-}
-func Test_getRawSQLs(t *testing.T) {
-	r1 := &Reader{
-		dbtype: ModeMySQL,
-	}
-	mysqltests := []struct {
-		queryType int
-		expSQLs   string
-	}{
-		{
-			queryType: TABLE,
-			expSQLs:   "Select * From `my_table`;",
-		},
-		{
-			queryType: COUNT,
-			expSQLs:   "Select Count(*) From `my_table`;",
-		},
-		{
-			queryType: DATABASE,
-			expSQLs:   "",
-		},
-	}
-
-	for _, test := range mysqltests {
-		sqls, err := r1.getRawSqls(test.queryType, "my_table")
-		assert.NoError(t, err)
-		assert.EqualValues(t, test.expSQLs, sqls)
-	}
-	r2 := Reader{
-		dbtype:   ModePostgreSQL,
-		dbSchema: "public",
-	}
-	pgtests := []struct {
-		queryType int
-		expSQLs   string
-	}{
-		{
-			queryType: TABLE,
-			expSQLs:   "Select * From \"public\".\"my_table\";",
-		},
-		{
-			queryType: COUNT,
-			expSQLs:   "Select Count(*) From \"public\".\"my_table\";",
-		},
-		{
-			queryType: DATABASE,
-			expSQLs:   "",
-		},
-	}
-	for _, test := range pgtests {
-		sqls, err := r2.getRawSqls(test.queryType, "my_table")
-		assert.NoError(t, err)
-		assert.EqualValues(t, test.expSQLs, sqls)
-	}
-	r3 := Reader{
-		dbtype:   ModeMSSQL,
-		dbSchema: "dbo",
-	}
-	mssqltests := []struct {
-		queryType int
-		expSQLs   string
-	}{
-		{
-			queryType: TABLE,
-			expSQLs:   "Select * From \"dbo\".\"my_table\";",
-		},
-		{
-			queryType: COUNT,
-			expSQLs:   "Select Count(*) From \"dbo\".\"my_table\";",
-		},
-		{
-			queryType: DATABASE,
-			expSQLs:   "",
-		},
-	}
-	for _, test := range mssqltests {
-		sqls, err := r3.getRawSqls(test.queryType, "my_table")
-		assert.NoError(t, err)
-		assert.EqualValues(t, test.expSQLs, sqls)
-	}
-
-}
-
-func Test_getConnectStr(t *testing.T) {
-	now := time.Date(2009, time.November, 10, 23, 0, 0, 0, time.UTC)
-
-	mr := &Reader{
-		database:    "Test_getCheckConnectStr",
-		dbtype:      "mysql",
-		historyAll:  true,
-		rawTable:    "*",
-		rawDatabase: "*",
-		datasource:  "root:@tcp(127.0.0.1:3306)",
-		encoder:     "utf8",
-	}
-	connectStr, err := mr.getConnectStr("", now)
-	assert.NoError(t, err)
-	expectMySql := "root:@tcp(127.0.0.1:3306)/?charset=utf8"
-	assert.Equal(t, expectMySql, connectStr)
-
-	mrMssql := &Reader{
-		database:    "Test_getCheckConnectStr",
-		dbtype:      "mssql",
-		rawTable:    "*",
-		datasource:  `server=localhost\SQLExpress;user id=sa;password=PassWord;port=1433`,
-		encoder:     "utf8",
-		rawDatabase: "Test_getCheckConnectStr",
-	}
-	connectStr, err = mrMssql.getConnectStr("", now)
-	assert.NoError(t, err)
-	expectMsSql := `server=localhost\SQLExpress;user id=sa;password=PassWord;port=1433;database=Test_getCheckConnectStr`
-	assert.Equal(t, expectMsSql, connectStr)
-
-	mrPostgressql := &Reader{
-		database:    "Test_getCheckConnectStr",
-		dbtype:      "postgres",
-		rawTable:    "*",
-		datasource:  "host=localhost port=5432 connect_timeout=10 user=pqgotest password=123456 sslmode=disable",
-		encoder:     "utf8",
-		rawDatabase: "Test_getCheckConnectStr",
-	}
-	connectStr, err = mrPostgressql.getConnectStr("", now)
-	assert.NoError(t, err)
-	expectPostgresSql := "host=localhost port=5432 connect_timeout=10 user=pqgotest password=123456 sslmode=disable dbname=Test_getCheckConnectStr"
-	assert.Equal(t, expectPostgresSql, connectStr)
-}
-
-func Test_WriteRecordsFile(t *testing.T) {
-	meta, err := reader.NewMeta(MetaDir, MetaDir, "mysql", "logpath", "", 7)
-	if err != nil {
-		t.Error(err)
-	}
-	defer os.RemoveAll(MetaDir)
-
-	err = WriteRecordsFile(meta.DoneFilePath, getContent(readRecords))
-	assert.NoError(t, err)
-}
-
-func Test_restoreRecordsFile(t *testing.T) {
-	meta, err := reader.NewMeta(MetaDir, MetaDir, "mysql", "logpath", "", 7)
-	if err != nil {
-		t.Error(err)
-	}
-	defer os.RemoveAll(MetaDir)
-
-	tests := []struct {
-		set                  DBRecords
-		expRes               DBRecords
-		expOmitDoneDBRecords bool
-	}{
-		{
-			set:                  DBRecords{},
-			expRes:               nil,
-			expOmitDoneDBRecords: true,
-		},
-		{
-			set:                  readRecords,
-			expRes:               readRecords,
-			expOmitDoneDBRecords: false,
-		},
-		{
-			set: DBRecords{
-				"db1": TableRecords{
-					Table: map[string]TableInfo{
-						"db1_tb1":  {size: -1, offset: -1},
-						"db1_tb10": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-				"db4": TableRecords{
-					Table: map[string]TableInfo{
-						"db4_tb10": TableInfo{size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-			},
-			expRes: DBRecords{
-				"db1": TableRecords{
-					Table: map[string]TableInfo{
-						"db1_tb1":  {size: -1, offset: -1},
-						"db1_tb2":  {size: -1, offset: -1},
-						"db1_tb3":  {size: -1, offset: -1},
-						"db1_tb10": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-				"db2": TableRecords{
-					Table: map[string]TableInfo{
-						"db2_tb1": {size: -1, offset: -1},
-						"db2_tb2": {size: -1, offset: -1},
-						"db2_tb3": {size: -1, offset: -1},
-						"db2_tb4": {size: -1, offset: -1},
-						"db2_tb5": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-				"db3": TableRecords{
-					Table: map[string]TableInfo{
-						"db3_tb1": {size: -1, offset: -1},
-						"db3_tb2": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-				"db4": TableRecords{
-					Table: map[string]TableInfo{
-						"db4_tb10": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-			},
-			expOmitDoneDBRecords: false,
-		},
-		{
-			set: DBRecords{
-				"db1": TableRecords{
-					Table: map[string]TableInfo{
-						"db1_tb1":  {size: -1, offset: -1},
-						"db1_tb10": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-				"db4": TableRecords{
-					Table: map[string]TableInfo{
-						"db4_tb10": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-			},
-			expRes: DBRecords{
-				"db1": TableRecords{
-					Table: map[string]TableInfo{
-						"db1_tb1":  {size: -1, offset: -1},
-						"db1_tb2":  {size: -1, offset: -1},
-						"db1_tb3":  {size: -1, offset: -1},
-						"db1_tb10": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-				"db2": TableRecords{
-					Table: map[string]TableInfo{
-						"db2_tb1": {size: -1, offset: -1},
-						"db2_tb2": {size: -1, offset: -1},
-						"db2_tb3": {size: -1, offset: -1},
-						"db2_tb4": {size: -1, offset: -1},
-						"db2_tb5": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-				"db3": TableRecords{
-					Table: map[string]TableInfo{
-						"db3_tb1": {size: -1, offset: -1},
-						"db3_tb2": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-				"db4": TableRecords{
-					Table: map[string]TableInfo{
-						"db4_tb10": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-			},
-			expOmitDoneDBRecords: false,
-		},
-		{
-			set: DBRecords{
-				"db1": TableRecords{
-					Table: map[string]TableInfo{
-						"db1_tb10": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-				"db3": TableRecords{
-					Table: map[string]TableInfo{
-						"db3_tb10": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-			},
-			expRes: DBRecords{
-				"db1": TableRecords{
-					Table: map[string]TableInfo{
-						"db1_tb1":  {size: -1, offset: -1},
-						"db1_tb2":  {size: -1, offset: -1},
-						"db1_tb3":  {size: -1, offset: -1},
-						"db1_tb10": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-				"db2": TableRecords{
-					Table: map[string]TableInfo{
-						"db2_tb1": {size: -1, offset: -1},
-						"db2_tb2": {size: -1, offset: -1},
-						"db2_tb3": {size: -1, offset: -1},
-						"db2_tb4": {size: -1, offset: -1},
-						"db2_tb5": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-				"db3": TableRecords{
-					Table: map[string]TableInfo{
-						"db3_tb1":  {size: -1, offset: -1},
-						"db3_tb2":  {size: -1, offset: -1},
-						"db3_tb10": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-				"db4": TableRecords{
-					Table: map[string]TableInfo{
-						"db4_tb10": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-			},
-			expOmitDoneDBRecords: false,
-		},
-		{
-			set: readRecords,
-			expRes: DBRecords{
-				"db1": TableRecords{
-					Table: map[string]TableInfo{
-						"db1_tb1":  {size: -1, offset: -1},
-						"db1_tb2":  {size: -1, offset: -1},
-						"db1_tb3":  {size: -1, offset: -1},
-						"db1_tb10": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-				"db2": TableRecords{
-					Table: map[string]TableInfo{
-						"db2_tb1": {size: -1, offset: -1},
-						"db2_tb2": {size: -1, offset: -1},
-						"db2_tb3": {size: -1, offset: -1},
-						"db2_tb4": {size: -1, offset: -1},
-						"db2_tb5": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-				"db3": TableRecords{
-					Table: map[string]TableInfo{
-						"db3_tb1":  {size: -1, offset: -1},
-						"db3_tb2":  {size: -1, offset: -1},
-						"db3_tb10": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-				"db4": TableRecords{
-					Table: map[string]TableInfo{
-						"db4_tb10": {size: -1, offset: -1},
-					},
-					mutex: sync.RWMutex{},
-				},
-			},
-			expOmitDoneDBRecords: false,
-		},
-	}
-
-	for _, test := range tests {
-		err = WriteRecordsFile(meta.DoneFilePath, getContent(test.set))
-		assert.NoError(t, err)
-
-		var records SyncDBRecords
-		_, _, omitDoneDBRecords := records.restoreRecordsFile(meta)
-		assert.EqualValues(t, test.expOmitDoneDBRecords, omitDoneDBRecords)
-		assert.EqualValues(t, test.expRes, records.GetDBRecords())
-	}
-
-}
-
-func getContent(readRecords DBRecords) string {
-	now := time.Now().String()
-	var all string
-	for database, tablesRecord := range readRecords {
-		var tablesRecordStr string
-		for table, tableInfo := range tablesRecord.GetTable() {
-			tablesRecordStr += table + "," +
-				strconv.FormatInt(tableInfo.size, 10) + "," +
-				strconv.FormatInt(tableInfo.size, 10) + "," +
-				now + "@"
-		}
-		all += database + sqlOffsetConnector + tablesRecordStr + "\n"
-	}
-
-	return all
-}
-
-func getMeta(metaDir string) (*reader.Meta, error) {
-	logkitConf := conf.MapConf{
-		KeyMetaPath: metaDir,
-		KeyFileDone: metaDir,
-		KeyMode:     ModeMySQL,
-	}
-	return reader.NewMetaWithConf(logkitConf)
-}
-
-type DataTest struct {
-	database    string
-	createTable []string
-	insertData  []string
-}
-
-type CronInfo struct {
-	cron           bool
-	second         string
-	notExecOnStart bool
-}
-
-var (
-	dbSource   = "root:@tcp(127.0.0.1:3306)"
-	pgDbSource = "host=127.0.0.1 port=5432 connect_timeout=10 user=postgres password=lyt  sslmode=disable"
-	connectStr = dbSource + "/?charset=gbk"
-	now        = time.Now()
-	year       = getDateStr(now.Year())
-	month      = getDateStr(int(now.Month()))
-	day        = getDateStr(now.Day())
-
-	databasesTest = []DataTest{
-		{
-			database:    "Test_MySql20180510",
-			createTable: []string{"CREATE TABLE runoob_tbl20180510est(runoob_id INT NOT NULL AUTO_INCREMENT,runoob_title VARCHAR(100) NOT NULL,runoob_author VARCHAR(40) NOT NULL,submission_date DATE,PRIMARY KEY ( runoob_id ))ENGINE=InnoDB DEFAULT CHARSET=utf8;"},
-			insertData:  []string{"INSERT INTO runoob_tbl20180510est (runoob_title, runoob_author, submission_date) VALUES (\"学习 mysql\", \"教程\", NOW());"},
-		},
-		{
-			database:    "Test_MySql20170610",
-			createTable: []string{"CREATE TABLE runoob_tbl20170610est(runoob_id INT NOT NULL AUTO_INCREMENT,runoob_title VARCHAR(100) NOT NULL,runoob_author VARCHAR(40) NOT NULL,submission_date DATE,PRIMARY KEY ( runoob_id ))ENGINE=InnoDB DEFAULT CHARSET=utf8;"},
-			insertData:  []string{"INSERT INTO runoob_tbl20170610est (runoob_title, runoob_author, submission_date) VALUES (\"学习 mysql\", \"教程\", NOW());"},
-		},
-		{
-			database:    "Test_MySql20171210",
-			createTable: []string{"CREATE TABLE runoob_tbl20171210est(runoob_id INT NOT NULL AUTO_INCREMENT,runoob_title VARCHAR(100) NOT NULL,runoob_author VARCHAR(40) NOT NULL,submission_date DATE,PRIMARY KEY ( runoob_id ))ENGINE=InnoDB DEFAULT CHARSET=utf8;"},
-			insertData:  []string{"INSERT INTO runoob_tbl20171210est (runoob_title, runoob_author, submission_date) VALUES (\"学习 mysql\", \"教程\", NOW());"},
-		},
-		{
-			database:    "Test_MySql20170910",
-			createTable: []string{"CREATE TABLE runoob_tbl20170910est(runoob_id INT NOT NULL AUTO_INCREMENT,runoob_title VARCHAR(100) NOT NULL,runoob_author VARCHAR(40) NOT NULL,submission_date DATE,PRIMARY KEY ( runoob_id ))ENGINE=InnoDB DEFAULT CHARSET=utf8;"},
-			insertData:  []string{"INSERT INTO runoob_tbl20170910est (runoob_title, runoob_author, submission_date) VALUES (\"学习 mysql\", \"教程\", NOW());"},
-		},
-		{
-			database:    "Test_MySql20180110",
-			createTable: []string{"CREATE TABLE runoob_tbl20180110est(runoob_id INT NOT NULL AUTO_INCREMENT,runoob_title VARCHAR(100) NOT NULL,runoob_author VARCHAR(40) NOT NULL,submission_date DATE,PRIMARY KEY ( runoob_id ))ENGINE=InnoDB DEFAULT CHARSET=utf8;"},
-			insertData:  []string{"INSERT INTO runoob_tbl20180110est (runoob_title, runoob_author, submission_date) VALUES (\"学习 mysql\", \"教程\", NOW());"},
-		},
-	}
-	dayDataTestsLine = len(databasesTest)
-	todayDataTests   = []DataTest{
-		{
-			"Test_MySql" + year + month + day,
-			[]string{"CREATE TABLE runoob_tbl" + year + month + day + "est(runoob_id INT NOT NULL AUTO_INCREMENT,runoob_title VARCHAR(100) NOT NULL,runoob_author VARCHAR(40) NOT NULL,submission_date DATE,PRIMARY KEY ( runoob_id ))ENGINE=InnoDB DEFAULT CHARSET=utf8;"},
-			[]string{"INSERT INTO runoob_tbl" + year + month + day + "est (runoob_title, runoob_author, submission_date) VALUES (\"学习 mysql\", \"教程\", NOW());"},
-		},
-	}
-	pgDatabaseTest = DataTest{
-		database:    "test_postgresdb",
-		createTable: []string{"CREATE TABLE person (id int4,name varchar,age int4,salary float4,delete  bool,create_time timestamp(6))WITH (OIDS=FALSE);"},
-		insertData: []string{
-			"INSERT INTO person VALUES ('1', '小王', null, '5000.3', 't', '2017-09-04 11:26:17');",  //test integer null
-			"INSERT INTO person VALUES ('2', '小明', '28',  null, 'f', '2018-03-20 11:22:17');",     //test float  null
-			"INSERT INTO person VALUES ('3', '小张', '28', '5000.5', null, '2018-10-10 11:23:17');", //test bool null
-		},
-	}
-	todayDataTestsLine = len(todayDataTests)
-)
-
-func TestMySql(t *testing.T) {
-	databasesTest = append(databasesTest, todayDataTests...)
-	if err := prepareMysql(); err != nil {
-		t.Errorf("prepare mysql database failed: %v", err)
-	}
-	defer func() {
-		if err := cleanMysql(); err != nil {
-			t.Errorf("clean mysql database failed: %v", err)
-		}
-	}()
-	expectData := Data{
-		"runoob_id":       int64(1),
-		"runoob_title":    "学习 mysql",
-		"runoob_author":   "教程",
-		"submission_date": year + "-" + month + "-" + day,
-	}
-
-	runnerName := "mrOnce"
-	err := os.Setenv("TestMySql_dbSource", dbSource)
-	assert.NoError(t, err)
-	defer os.Unsetenv("TestMySql_dbSource")
-	mr, err := reader.NewReader(conf.MapConf{
-		"mysql_database":     "Test_MySql20180510",
-		"mysql_table":        "runoob_tbl20180510est",
-		"mysql_limit_batch":  "100",
-		"mysql_history_all":  "true",
-		"mode":               "mysql",
-		"mysql_exec_onstart": "true",
-		"encoding":           "gbk",
-		"mysql_datasource":   "${TestMySql_dbSource}",
-		"meta_path":          path.Join(MetaDir, runnerName),
-		"file_done":          path.Join(MetaDir, runnerName),
-		"runner_name":        runnerName,
-	}, true)
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.RemoveAll(MetaDir)
-	r, ok := mr.(reader.DataReader)
-	if !ok {
-		t.Error("mysql read should have readdata interface")
-	}
-	assert.NoError(t, mr.(*Reader).Start())
-
-	dataLine := 0
-	before := time.Now()
-	for !batchTimeout(before, 2) {
-		data, bytes, err := r.ReadData()
-		if err != nil {
-			t.Error(err)
-		}
-		if len(data) <= 0 {
-			continue
-		}
-		assert.Equal(t, int64(36), bytes)
-		assert.Equal(t, expectData, data)
-		dataLine++
-	}
-	assert.Equal(t, 1, dataLine)
-
-	// test exec on start
-	runnerName = "mr"
-	mr, err = getMySqlReader(false, false, false, runnerName, CronInfo{})
-	assert.NoError(t, err)
-	r, ok = mr.(reader.DataReader)
-	if !ok {
-		t.Error("mysql read should have readdata interface")
-	}
-	assert.NoError(t, mr.(*Reader).Start())
-
-	dataLine = 0
-	before = time.Now()
-	for !batchTimeout(before, 2) {
-		data, bytes, err := r.ReadData()
-		if err != nil {
-			t.Error(err)
-		}
-		if len(data) <= 0 {
-			continue
-		}
-		assert.Equal(t, int64(36), bytes)
-		assert.Equal(t, expectData, data)
-		dataLine++
-	}
-	assert.Equal(t, todayDataTestsLine, dataLine)
-	mr.SyncMeta()
-
-	// test sync records
-	dataLine = 0
-	before = time.Now()
-	for !batchTimeout(before, 2) {
-		data, bytes, err := r.ReadData()
-		if err != nil {
-			t.Error(err)
-		}
-		if len(data) <= 0 {
-			continue
-		}
-		assert.Equal(t, int64(36), bytes)
-		assert.Equal(t, expectData, data)
-		dataLine++
-	}
-	assert.Equal(t, 0, dataLine)
-	mr.SyncMeta()
-	mr.Close()
-
-	// test exec on start, sql not empty
-	runnerName = "mrRawSql"
-	mrRawSql, err := getMySqlReader(false, true, false, runnerName, CronInfo{})
-	assert.NoError(t, err)
-	mrRawSqlData, ok := mrRawSql.(reader.DataReader)
-	if !ok {
-		t.Error("mysql read should have readdata interface")
-	}
-	assert.NoError(t, mrRawSql.(*Reader).Start())
-
-	dataLine = 0
-	before = time.Now()
-	for !batchTimeout(before, 2) {
-		data, bytes, err := mrRawSqlData.ReadData()
-		if err != nil {
-			t.Error(err)
-		}
-		if len(data) <= 0 {
-			continue
-		}
-		assert.Equal(t, int64(36), bytes)
-		assert.Equal(t, expectData, data)
-		dataLine++
-	}
-	assert.Equal(t, todayDataTestsLine, dataLine)
-	mrRawSql.SyncMeta()
-
-	// no sync records when raw sql is not empty
-	dataLine = 0
-	before = time.Now()
-	for !batchTimeout(before, 2) {
-		data, bytes, err := mrRawSqlData.ReadData()
-		if err != nil {
-			t.Error(err)
-		}
-		if len(data) <= 0 {
-			continue
-		}
-		assert.Equal(t, int64(36), bytes)
-		assert.Equal(t, expectData, data)
-		dataLine++
-	}
-	assert.Equal(t, 0, dataLine)
-	mrRawSql.SyncMeta()
-	mrRawSql.Close()
-
-	// test history all
-	runnerName = "mrHistoryAll"
-	mrHistoryAll, err := getMySqlReader(true, false, false, runnerName, CronInfo{})
-	assert.NoError(t, err)
-	mrHistoryAllData, ok := mrHistoryAll.(reader.DataReader)
-	if !ok {
-		t.Error("mysql read should have readdata interface")
-	}
-	assert.NoError(t, mrHistoryAll.(*Reader).Start())
-
-	dataLine = 0
-	before = time.Now()
-	for !batchTimeout(before, 2) {
-		data, bytes, err := mrHistoryAllData.ReadData()
-		if err != nil {
-			t.Error(err)
-		}
-		if len(data) <= 0 {
-			continue
-		}
-		assert.Equal(t, int64(36), bytes)
-		assert.Equal(t, expectData, data)
-		dataLine++
-	}
-	assert.Equal(t, dayDataTestsLine+todayDataTestsLine, dataLine)
-	mrHistoryAll.SyncMeta()
-	mrHistoryAll.Close()
-
-	// test file done in meta dir
-	mrHistoryAll2, err := getMySqlReader(true, false, false, runnerName, CronInfo{})
-	assert.NoError(t, err)
-	mrHistoryAllData2, ok := mrHistoryAll2.(reader.DataReader)
-	if !ok {
-		t.Error("mysql read should have readdata interface")
-	}
-	dataLine = 0
-	before = time.Now()
-	for !batchTimeout(before, 2) {
-		data, _, err := mrHistoryAllData2.ReadData()
-		if err != nil {
-			t.Error(err)
-		}
-		if len(data) <= 0 {
-			continue
-		}
-		dataLine++
-	}
-	assert.Equal(t, 0, dataLine)
-	mrHistoryAll2.SyncMeta()
-	mrHistoryAll2.Close()
-
-	// test cron
-	minDataTestsLine, secondAdd3, err := setMinute(time.Now())
-	if err != nil {
-		t.Errorf("prepare mysql database failed: %v", err)
-	}
-	// cron task, not exec on start
-	runnerName = "mrCron"
-	mrCron, err := getMySqlReader(false, false, false, runnerName, CronInfo{true, secondAdd3, true})
-	assert.NoError(t, err)
-	mrCronData, ok := mrCron.(reader.DataReader)
-	if !ok {
-		t.Error("mysql read should have readdata interface")
-	}
-	assert.NoError(t, mrCron.(*Reader).Start())
-
-	dataLine = 0
-	before = time.Now()
-	log.Infof("before: %v", before)
-	for !batchTimeout(before, 5) {
-		data, bytes, err := mrCronData.ReadData()
-		if err != nil {
-			t.Error(err)
-		}
-		if len(data) <= 0 {
-			continue
-		}
-		assert.Equal(t, int64(36), bytes)
-		assert.Equal(t, expectData, data)
-		dataLine++
-	}
-	assert.Equal(t, minDataTestsLine, dataLine)
-	mrCron.SyncMeta()
-	mrCron.Close()
-
-	// cron task, exec on start
-	now := time.Now()
-	minDataTestsLine, secondAdd3, err = setMinute(now)
-	if err != nil {
-		t.Errorf("prepare mysql database failed: %v", err)
-	}
-	if now.Second() >= 57 {
-		minDataTestsLine++
-	}
-	runnerName = "mrCronExecOnStart"
-	mrCronExecOnStart, err := getMySqlReader(false, false, false, runnerName, CronInfo{true, secondAdd3, false})
-	assert.NoError(t, err)
-	mrCronExecOnStartData, ok := mrCronExecOnStart.(reader.DataReader)
-	if !ok {
-		t.Error("mysql read should have readdata interface")
-	}
-	assert.NoError(t, mrCronExecOnStart.(*Reader).Start())
-
-	dataLine = 0
-	before = time.Now()
-	log.Infof("before: %v", before)
-	for !batchTimeout(before, 5) {
-		data, bytes, err := mrCronExecOnStartData.ReadData()
-		if err != nil {
-			t.Error(err)
-		}
-		if len(data) <= 0 {
-			continue
-		}
-		assert.Equal(t, int64(36), bytes)
-		assert.Equal(t, expectData, data)
-		dataLine++
-	}
-	assert.Equal(t, minDataTestsLine, dataLine)
-	mrCronExecOnStart.SyncMeta()
-	mrCronExecOnStart.Close()
-
-	mrCronExecOnStart2, err := getMySqlReader(false, false, false, runnerName, CronInfo{true, secondAdd3, false})
-	assert.NoError(t, err)
-	mrCronExecOnStartData2, ok := mrCronExecOnStart2.(reader.DataReader)
-	if !ok {
-		t.Error("mysql read should have readdata interface")
-	}
-	assert.NoError(t, mrCronExecOnStart2.(*Reader).Start())
-
-	dataLine = 0
-	before = time.Now()
-	log.Infof("before: %v", before)
-	for !batchTimeout(before, 5) {
-		data, bytes, err := mrCronExecOnStartData2.ReadData()
-		if err != nil {
-			t.Error(err)
-		}
-		if len(data) <= 0 {
-			continue
-		}
-		assert.Equal(t, int64(36), bytes)
-		assert.Equal(t, expectData, data)
-		dataLine++
-	}
-	assert.Equal(t, 0, dataLine)
-	mrCronExecOnStart2.SyncMeta()
-	mrCronExecOnStart2.Close()
-
-	// cron task, exec on start
-	minDataTestsLine, _, err = setSecond()
-	if err != nil {
-		t.Errorf("prepare mysql database failed: %v", err)
-	}
-	runnerName = "mrLoopcOnStart"
-	mrLoopOnStart, err := getMySqlReader(false, false, true, runnerName, CronInfo{false, "", false})
-	assert.NoError(t, err)
-	mrLoopOnStartData, ok := mrLoopOnStart.(reader.DataReader)
-	if !ok {
-		t.Error("mysql read should have readdata interface")
-	}
-	assert.NoError(t, mrLoopOnStart.(*Reader).Start())
-
-	dataLine = 0
-	before = time.Now()
-	log.Infof("before: %v", before)
-	for !batchTimeout(before, 5) {
-		data, bytes, err := mrLoopOnStartData.ReadData()
-		if err != nil {
-			t.Error(err)
-		}
-		if len(data) <= 0 {
-			continue
-		}
-		assert.Equal(t, int64(36), bytes)
-		assert.Equal(t, expectData, data)
-		dataLine++
-	}
-	assert.Equal(t, minDataTestsLine, dataLine)
-	mrLoopOnStart.SyncMeta()
-
-	meta, err := getMeta(path.Join(MetaDir, runnerName))
-	assert.NoError(t, err)
-	var doneRecords = SyncDBRecords{
-		mutex: sync.RWMutex{},
-	}
-	lastDB, lastTable, omitDoneFile := doneRecords.restoreRecordsFile(meta)
-	assert.False(t, omitDoneFile)
-	assert.Equal(t, 1, len(doneRecords.records))
-	expectDB := "Test_MySql" + year + month + day
-	tableRecords := doneRecords.records.GetTableRecords(expectDB)
-	assert.Equal(t, 2, len(tableRecords.GetTable()))
-	assert.Equal(t, expectDB, lastDB)
-	assert.NotEmpty(t, lastTable)
-	mrLoopOnStart.Close()
-}
-
-func getMySqlReader(historyAll, rawsql, loop bool, runnerName string, cronInfo CronInfo) (reader.Reader, error) {
-	readerConf := conf.MapConf{
-		"mysql_database":     "Test_MySql@(YYYY)@(MM)@(DD)",
-		"mysql_table":        "runoob_tbl@(YYYY)@(MM)@(DD)est",
-		"mysql_limit_batch":  "100",
-		"mysql_history_all":  "false",
-		"mode":               "mysql",
-		"mysql_exec_onstart": "true",
-		"encoding":           "gbk",
-		"mysql_datasource":   dbSource,
-		"meta_path":          path.Join(MetaDir, runnerName),
-		"file_done":          path.Join(MetaDir, runnerName),
-		"runner_name":        runnerName,
-	}
-	if rawsql {
-		readerConf["mysql_table"] = ""
-		readerConf["mysql_sql"] = "select * from runoob_tbl" + year + month + day + "est"
-	}
-	if historyAll {
-		readerConf["mysql_history_all"] = "true"
-	}
-	if cronInfo.cron {
-		readerConf["mysql_cron"] = cronInfo.second + " * * * * *"
-		readerConf["mysql_database"] = readerConf["mysql_database"] + "@(mm)"
-		readerConf["mysql_table"] = "runoob_tbl@(YYYY)@(MM)@(DD)" + "@(mm)" + "est"
-	}
-	if cronInfo.notExecOnStart {
-		readerConf["mysql_exec_onstart"] = "false"
-	}
-	if loop {
-		readerConf["mysql_cron"] = "loop 1s"
-		readerConf["mysql_table"] = "runoob_tbl@(YYYY)@(MM)@(DD)@(ss)est"
-	}
-	mr, err := reader.NewReader(readerConf, true)
-	if err != nil {
-		return nil, err
-	}
-	return mr, nil
-}
-
-func prepareMysql() error {
-	db, err := openSql("mysql", connectStr)
-	if err != nil {
-		return err
-	}
-
-	defer db.Close()
-	if err = db.Ping(); err != nil {
-		return err
-	}
-
-	for _, dbInfo := range databasesTest {
-		_, err = db.Query("DROP DATABASE IF EXISTS " + dbInfo.database)
-		if err != nil {
-			return err
-		}
-
-		_, err = db.Exec("CREATE DATABASE " + dbInfo.database)
-		if err != nil {
-			return err
-		}
-
-		_, err = db.Exec("USE " + dbInfo.database)
-		if err != nil {
-			return err
-		}
-
-		for _, createTable := range dbInfo.createTable {
-			_, err = db.Exec(createTable)
-			if err != nil {
-				return err
-			}
-		}
-
-		for _, data := range dbInfo.insertData {
-			_, err = db.Exec(data)
-			if err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
-func cleanMysql() error {
-	db, err := openSql("mysql", connectStr)
-	if err != nil {
-		return err
-	}
-	for _, dbInfo := range databasesTest {
-		_, err := db.Query("DROP DATABASE IF EXISTS " + dbInfo.database)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func getDateStr(date int) string {
-	dateStr := strconv.Itoa(date)
-	if date < 10 {
-		return "0" + dateStr
-	}
-	return dateStr
-}
-
-func batchTimeout(before time.Time, interval float64) bool {
-	// 超过最长的发送间隔
-	if time.Now().Sub(before).Seconds() >= interval {
-		return true
-	}
-
-	return false
-}
-
-func setMinute(nowCron time.Time) (int, string, error) {
-	var (
-		secondAdd3 = getDateStr((nowCron.Second() + 3) % 60)
-		minute     = getDateStr(nowCron.Minute())
-	)
-	if nowCron.Second() >= 57 {
-		minute = getDateStr(nowCron.Minute() + 1)
-	}
-	var minDataTests = []DataTest{
-		{
-			"Test_MySql" + year + month + day + minute,
-			[]string{"CREATE TABLE runoob_tbl" + year + month + day + minute + "est(runoob_id INT NOT NULL AUTO_INCREMENT,runoob_title VARCHAR(100) NOT NULL,runoob_author VARCHAR(40) NOT NULL,submission_date DATE,PRIMARY KEY ( runoob_id ))ENGINE=InnoDB DEFAULT CHARSET=utf8;"},
-			[]string{"INSERT INTO runoob_tbl" + year + month + day + minute + "est (runoob_title, runoob_author, submission_date) VALUES (\"学习 mysql\", \"教程\", NOW());"},
-		},
-	}
-	minDataTestsLine := len(minDataTests)
-	log.Infof("time now cron: %v, minute: %v, secondAdd3: %v", nowCron, minute, secondAdd3)
-	databasesTest = append(databasesTest, minDataTests...)
-	if err := cleanMysql(); err != nil {
-		return minDataTestsLine, secondAdd3, err
-	}
-	if err := prepareMysql(); err != nil {
-		return minDataTestsLine, secondAdd3, err
-	}
-	return minDataTestsLine, secondAdd3, nil
-}
-
-func setSecond() (int, string, error) {
-	var (
-		nowCron    = time.Now()
-		secondAdd2 = getDateStr((nowCron.Second() + 2) % 60)
-		secondAdd4 = getDateStr((nowCron.Second() + 4) % 60)
-		minute     = getDateStr(nowCron.Minute())
-	)
-	var minDataTests = []DataTest{
-		{
-			"Test_MySql" + year + month + day,
-			[]string{"CREATE TABLE runoob_tbl" + year + month + day + secondAdd2 + "est(runoob_id INT NOT NULL AUTO_INCREMENT,runoob_title VARCHAR(100) NOT NULL,runoob_author VARCHAR(40) NOT NULL,submission_date DATE,PRIMARY KEY ( runoob_id ))ENGINE=InnoDB DEFAULT CHARSET=utf8;",
-				"CREATE TABLE runoob_tbl" + year + month + day + secondAdd4 + "est(runoob_id INT NOT NULL AUTO_INCREMENT,runoob_title VARCHAR(100) NOT NULL,runoob_author VARCHAR(40) NOT NULL,submission_date DATE,PRIMARY KEY ( runoob_id ))ENGINE=InnoDB DEFAULT CHARSET=utf8;"},
-			[]string{"INSERT INTO runoob_tbl" + year + month + day + secondAdd2 + "est (runoob_title, runoob_author, submission_date) VALUES (\"学习 mysql\", \"教程\", NOW());",
-				"INSERT INTO runoob_tbl" + year + month + day + secondAdd4 + "est (runoob_title, runoob_author, submission_date) VALUES (\"学习 mysql\", \"教程\", NOW());"},
-		},
-	}
-	log.Infof("time now cron: %v, minute: %v, secondAdd2: %v, secondAdd4: %v", nowCron, minute, secondAdd2, secondAdd4)
-	databasesTest = append(databasesTest, minDataTests...)
-	if err := cleanMysql(); err != nil {
-		return 0, "", err
-	}
-	if err := prepareMysql(); err != nil {
-		return 0, "", err
-	}
-	return 2, "", nil
-}
-
-//for postgres
-func TestLocation(t *testing.T) {
-	tm1, _ := time.ParseInLocation(time.RFC3339, "2017-09-04T11:26:17Z", time.FixedZone("+0000", 0))
-	tm1 = time.Date(2017, 9, 4, 11, 26, 17, 0, time.FixedZone("+0000", 0))
-	fmt.Println(tm1.String())
-}
-
-func TestPostgres(t *testing.T) {
-	if err := preparePostgres(0); err != nil {
-		t.Fatalf("prepare postgres database failed: %v", err)
-	}
-	tm1 := time.Date(2017, 9, 4, 11, 26, 17, 0, time.FixedZone("+0000", 0))
-	tm2 := time.Date(2018, 3, 20, 11, 22, 17, 0, time.FixedZone("+0000", 0))
-	tm3 := time.Date(2018, 10, 10, 11, 23, 17, 0, time.FixedZone("+0000", 0))
-	expectDatas := []Data{
-		{
-			"id":          int64(1),
-			"name":        "小王",
-			"age":         int64(0),
-			"salary":      5000.2998046875,
-			"delete":      true,
-			"create_time": tm1,
-		},
-		{
-			"id":          int64(2),
-			"name":        "小明",
-			"age":         int64(28),
-			"salary":      float64(0),
-			"delete":      false,
-			"create_time": tm2,
-		},
-		{
-			"id":          int64(3),
-			"name":        "小张",
-			"age":         int64(28),
-			"salary":      5000.5,
-			"delete":      false,
-			"create_time": tm3,
-		},
-	}
-
-	runnerName := "TestPostgres"
-	mr, err := reader.NewReader(conf.MapConf{
-		"postgres_database":     pgDatabaseTest.database + "0",
-		"postgres_limit_batch":  "100",
-		"mode":                  "postgres",
-		"postgres_exec_onstart": "true",
-		"postgres_datasource":   pgDbSource,
-		"postgres_sql":          "select * from person",
-		"meta_path":             path.Join(MetaDir, runnerName),
-		"file_done":             path.Join(MetaDir, runnerName),
-		"runner_name":           runnerName,
-	}, true)
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.RemoveAll(MetaDir)
-	r, ok := mr.(reader.DataReader)
-	if !ok {
-		t.Error("postgres read should have readdata interface")
-	}
-	assert.NoError(t, mr.(*Reader).Start())
-
-	dataLine := 0
-	before := time.Now()
-	var actualData []Data
-	for !batchTimeout(before, 2) {
-		data, _, err := r.ReadData()
-		if err != nil {
-			t.Error(err)
-		}
-		if len(data) <= 0 {
-			continue
-		}
-		actualData = append(actualData, data)
-		dataLine++
-
-	}
-	assert.Equal(t, 3, dataLine)
-
-	for k, v := range actualData {
-		delete(expectDatas[k], "create_time")
-		delete(v, "create_time")
-		assert.Equal(t, expectDatas[k], v)
-	}
-}
-
-func TestPostgresWithOffset(t *testing.T) {
-	if err := preparePostgres(1); err != nil {
-		t.Fatalf("prepare postgres database failed: %v", err)
-	}
-	tm1 := time.Date(2017, 9, 4, 11, 26, 17, 0, time.FixedZone("+0000", 0))
-	tm2 := time.Date(2018, 3, 20, 11, 22, 17, 0, time.FixedZone("+0000", 0))
-	tm3 := time.Date(2018, 10, 10, 11, 23, 17, 0, time.FixedZone("+0000", 0))
-	expectDatas := []Data{
-		{
-			"id":          int64(1),
-			"name":        "小王",
-			"age":         int64(0),
-			"salary":      5000.2998046875,
-			"delete":      true,
-			"create_time": tm1,
-		},
-		{
-			"id":          int64(2),
-			"name":        "小明",
-			"age":         int64(28),
-			"salary":      float64(0),
-			"delete":      false,
-			"create_time": tm2,
-		},
-		{
-			"id":          int64(3),
-			"name":        "小张",
-			"age":         int64(28),
-			"salary":      5000.5,
-			"delete":      false,
-			"create_time": tm3,
-		},
-	}
-
-	os.MkdirAll(MetaDir+"/TestPostgres/", 0777)
-	err := ioutil.WriteFile("meta/TestPostgres/file.meta", []byte("select@*@from@person##2\t1\n"), 0777)
-	assert.NoError(t, err)
-
-	runnerName := "TestPostgres"
-	mr, err := reader.NewReader(conf.MapConf{
-		"postgres_database":     pgDatabaseTest.database + "1",
-		"postgres_limit_batch":  "100",
-		"mode":                  "postgres",
-		"postgres_offset_key":   "id",
-		"postgres_exec_onstart": "true",
-		"postgres_datasource":   pgDbSource,
-		"postgres_sql":          "select * from person",
-		"meta_path":             path.Join(MetaDir, runnerName),
-		"file_done":             path.Join(MetaDir, runnerName),
-		"runner_name":           runnerName,
-	}, true)
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.RemoveAll(MetaDir)
-	r, ok := mr.(reader.DataReader)
-	if !ok {
-		t.Error("postgres read should have readdata interface")
-	}
-	assert.NoError(t, mr.(*Reader).Start())
-
-	dataLine := 0
-	before := time.Now()
-	var actualData []Data
-	for !batchTimeout(before, 2) {
-		data, _, err := r.ReadData()
-		if err != nil {
-			t.Error(err)
-		}
-		if len(data) <= 0 {
-			continue
-		}
-		actualData = append(actualData, data)
-		dataLine++
-
-	}
-	assert.Equal(t, 2, dataLine)
-
-	for i := 0; i < len(actualData); i++ {
-		delete(expectDatas[i+1], "create_time")
-		delete(actualData[i], "create_time")
-		assert.Equal(t, expectDatas[i+1], actualData[i])
-	}
-}
-
-func getPostgresDb(dbsource string) (db *sql.DB, err error) {
-	db, err = openSql("postgres", dbsource)
-	if err != nil {
-		return nil, err
-	}
-	if err = db.Ping(); err != nil {
-		return nil, err
-	}
-	return db, nil
-}
-
-func dropAndCreateDB(id int) error {
-	db, err := getPostgresDb(pgDbSource)
-	if err != nil {
-		return err
-	}
-	defer db.Close()
-	_, err = db.Query("DROP DATABASE IF EXISTS " + pgDatabaseTest.database + strconv.Itoa(id))
-	if err != nil {
-		return err
-	}
-	_, err = db.Exec("CREATE DATABASE " + pgDatabaseTest.database + strconv.Itoa(id))
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-//使用相同数据库会由于排他锁而无法访问 故加一个随机数以区分
-func preparePostgres(random int) error {
-	dropAndCreateDB(random)
-
-	db, err := getPostgresDb(pgDbSource + " dbname=" + pgDatabaseTest.database + strconv.Itoa(random))
-	if err != nil {
-		return err
-	}
-	for _, createTable := range pgDatabaseTest.createTable {
-		_, err = db.Exec(createTable)
-		if err != nil {
-			return err
-		}
-	}
-	for _, data := range pgDatabaseTest.insertData {
-		_, err = db.Exec(data)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func TestPostgresWithTimestampInt(t *testing.T) {
-	t.Parallel()
-	if err := dropAndCreateDB(2); err != nil {
-		t.Fatalf("prepare postgres database failed: %v", err)
-	}
-
-	os.MkdirAll(MetaDir+"/TestPostgresWithTimestamp/", 0777)
-	defer os.RemoveAll(MetaDir)
-	tablename := "testtime1"
-	runnerName := "TestPostgresWithTimestamp"
-	mr, err := reader.NewReader(conf.MapConf{
-		"postgres_database": pgDatabaseTest.database + "2",
-		"mode":              "postgres",
-		"postgres_timestamp_key":  "timestamp",
-		"postgres_exec_onstart":   "true",
-		"postgres_start_time":     "0",
-		"postgres_batch_intervel": "1000",
-		"postgres_timestamp_int":  "true",
-		"postgres_cron":           "loop 1s",
-		"postgres_datasource":     pgDbSource,
-		"postgres_sql":            "select * from " + tablename,
-		"meta_path":               path.Join(MetaDir, runnerName),
-		"file_done":               path.Join(MetaDir, runnerName),
-		"runner_name":             runnerName,
-	}, true)
-	if err != nil {
-		t.Fatal(err)
-	}
-	r, ok := mr.(reader.DataReader)
-	if !ok {
-		t.Error("postgres read should have readdata interface")
-	}
-	assert.NoError(t, mr.(*Reader).Start())
-	totalnum := 10000
-	go datagen.GeneratePostgresData(pgDbSource+" dbname="+pgDatabaseTest.database+"2", tablename, int64(totalnum), 100*time.Millisecond, time.Hour, time.Now().Add(-100*time.Hour))
-	dataLine := 0
-	before := time.Now()
-	var actualData []Data
-	for !batchTimeout(before, 60) {
-		data, _, err := r.ReadData()
-		if err != nil {
-			continue
-		}
-		if len(data) <= 0 {
-			continue
-		}
-		actualData = append(actualData, data)
-		dataLine++
-		if dataLine >= totalnum {
-			break
-		}
-	}
-	assert.Equal(t, totalnum, dataLine)
-
-}
-
-func TestPostgresWithTimestampS(t *testing.T) {
-	t.Parallel()
-	if err := dropAndCreateDB(3); err != nil {
-		t.Fatalf("prepare postgres database failed: %v", err)
-	}
-
-	os.MkdirAll(MetaDir+"/TestPostgresWithTimestampS/", 0777)
-	defer os.RemoveAll(MetaDir)
-	tablename := "testtime2"
-	runnerName := "TestPostgresWithTimestampS"
-	mr, err := reader.NewReader(conf.MapConf{
-		"postgres_database": pgDatabaseTest.database + "3",
-		"mode":              "postgres",
-		"postgres_timestamp_key":  "create_time",
-		"postgres_exec_onstart":   "true",
-		"postgres_start_time":     "2018-10-01 15:04:05",
-		"postgres_batch_intervel": "30m",
-		"postgres_cron":           "loop 1s",
-		"postgres_datasource":     pgDbSource,
-		"postgres_sql":            "select * from " + tablename,
-		"meta_path":               path.Join(MetaDir, runnerName),
-		"file_done":               path.Join(MetaDir, runnerName),
-		"runner_name":             runnerName,
-	}, true)
-	if err != nil {
-		t.Fatal(err)
-	}
-	r, ok := mr.(reader.DataReader)
-	if !ok {
-		t.Error("postgres read should have readdata interface")
-	}
-	assert.NoError(t, mr.(*Reader).Start())
-	totalnum := 10000
-	go datagen.GeneratePostgresData(pgDbSource+" dbname="+pgDatabaseTest.database+"3", tablename, int64(totalnum), 100*time.Millisecond, time.Hour, time.Now().Add(-100*time.Hour))
-	dataLine := 0
-	before := time.Now()
-	var actualData []Data
-	for !batchTimeout(before, 60) {
-		data, _, err := r.ReadData()
-		if err != nil {
-			continue
-		}
-		if len(data) <= 0 {
-			continue
-		}
-		actualData = append(actualData, data)
-		dataLine++
-		if dataLine >= totalnum {
-			break
-		}
-	}
-	assert.Equal(t, totalnum, dataLine)
-
-}
